requirements.txt
jupyter>=1.1.1
notebook>=7.5.3
ipykernel>=7.1.0
numpy<2.0.0
pandas>=3.0.0
matplotlib>=3.10.8
scikit-learn>=1.6.0
scikit-surprise>=1.1.3
joblib>=1.4.2
scipy>=1.15.0
uv>=0.5.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
python-jose[cryptography]>=3.5.0